requirements.txt
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
pytest-django==4.7.0
pytest-mock==3.12.0